MB-50909 Add support for comparing expected and actual number of items
Added necessary variables and logging for comparing how many items
are expected to be backed up and how many items had actually been
backed up. This is going to be useful when testing PiTR backups with
data that is generated using cbc-pillowfight since for some settings
it might not be able to generate enough mutations for each
granularity period.